# improve performance on SyntaxErrors